Finished OLA

+++ public/api.js
@@ -4,7 +4,6 @@ export async function createUser(obj) {
     headers: { "Content-Type": "application/json" },
     body: JSON.stringify(obj),
   });
-  console.log(response);
   document.querySelector("form").reset();
 
   return response;
@@ -19,13 +18,12 @@ export async function saveToDatabase(obj) {
   return response;
 }
 
-export const getDestinations = async () => {
+export async function getDestinations() {
   const response = await fetch("/api/destinations");
   const result = await response.json();
-  console.log(result);
 
   return result;
-};
+}
 
 export async function updateDestination(id) {
   const imagePath = document.querySelector("#image-upload").value;
@@ -40,22 +38,26 @@ export async function updateDestination(id) {
     imageData = response.image;
   }
 
-  const response = await (
-    await fetch(`/api/destination/${id}`, {
-      // TODO BEARER AUTH
-      method: "PUT",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        country: document.querySelector("#country").value,
-        title: document.querySelector("#title").value,
-        link: document.querySelector("#link").value,
-        arrivalDate: document.querySelector("#arrival-date").value,
-        departureDate: document.querySelector("#departure-date").value,
-        image: imageData,
-        description: document.querySelector("#description").value,
-      }),
-    })
-  ).json();
+  const request = await fetch(`/api/destination/${id}`, {
+    method: "PUT",
+    headers: {
+      "Content-Type": "application/json",
+      Authorization: `Bearer ${localStorage.getItem("jwtToken")}`,
+    },
+    body: JSON.stringify({
+      country: document.querySelector("#country").value,
+      title: document.querySelector("#title").value,
+      link: document.querySelector("#link").value,
+      arrivalDate: document.querySelector("#arrival-date").value,
+      departureDate: document.querySelector("#departure-date").value,
+      image: imageData,
+      description: document.querySelector("#description").value,
+    }),
+  });
+  if (request.status === 401)
+    return alert("Failed to update destination - Unauthorized");
+
+  const response = await request.json();
 
   if (response.status === "Successful") {
     window.location.href = "/";
@@ -65,13 +67,21 @@ export async function updateDestination(id) {
 }
 
 export async function deleteDestination(id) {
-  const response = await (
-    await fetch(`/api/destination/${id}`, {
-      // TODO BEARER AUTH
-      method: "DELETE",
-      headers: { "Content-Type": "application/json" },
-    })
-  ).json();
+  if (!confirm("Are you sure you would like to delete this destination?"))
+    return;
+
+  const request = await fetch(`/api/destination/${id}`, {
+    method: "DELETE",
+    headers: {
+      "Content-Type": "application/json",
+      Authorization: `Bearer ${localStorage.getItem("jwtToken")}`,
+    },
+  });
+
+  if (request.status === 401)
+    return alert("Failed to delete destination - Unauthorized");
+
+  const response = await request.json();
 
   if (response.status === "Successful") {
     window.location.href = "/";
